Reimplementation of PostAdapter with Data Binding Library
By using the Data Binding Library, the implemenation of a custom
adapter is simplified by maintaining the same performance gains
(or maybe even more) as with the ViewHolder pattern.

In some way, the data binding of a ListView item view works like a
"built-in" ViewHolder.

Instead of a manually created ViewHolder, an automatically
created data binding object is saved in the tag of each newly
created list item view from where it can be retrievd in future
recycling iterations. The components of the list item view are
then accessed through this data binding object.